+++ pages/api/getSmsLogMessageByUserId.js
@@ -9,6 +9,8 @@ const getSmsLogMessageByUserId = async (req, res) => {
   }
 
   let auth = [];
+  let authMessage = [];
+  let messages = [];
   
   let { data: users, error } = await supabase
     .from('sms_log_message')
@@ -23,13 +25,15 @@ const getSmsLogMessageByUserId = async (req, res) => {
       .select(`*, sms_log_message (*)`)
       .eq('email', authEmail)
       .eq('sms_log_message.to_phone_number', users[0].from_phone_number)
-  } 
-
-  const authMessage = auth.data[0].sms_log_message.map((d) => ({ ...d, isOwnMessage: true }));
-
-  const messages = [...users, ...authMessage].sort((a, b) => new Date(a.created_at).getTime() - new Date(b.created_at).getTime());
+  }
+  
+  if (auth.data && auth.data.length > 0) {
+     authMessage = auth.data[0].sms_log_message.map((d) => ({ ...d, isOwnMessage: true }));
+     messages = [...users, ...authMessage].sort((a, b) => new Date(a.created_at).getTime() - new Date(b.created_at).getTime());
+  }
 
   if (error || auth.error) return res.status(401).json({ error: error.message })
+
   return res.status(200).json(messages)
 }
